+++ src/constants.ts
@@ -1,5 +1,6 @@
 // Copyright (c) 2021 Visiosto oy
 // Licensed under the MIT License
 
-// eslint-disable-next-line import/prefer-default-export
 export const initialColorModeCSSProp = '--initial-color-mode';
+
+export const formClassErrorMessage = 'error-message';