+++ src/gatsby/anchor-link/stripHashedLocation.js
@@ -1,9 +1,0 @@
-// Copyright (c) 2021 Visiosto oy
-// Licensed under the MIT License
-
-// Based on code by Chase Ohlson.
-// Original code is available at https://github.com/brohlson/gatsby-plugin-anchor-links.
-
-export default function stripHashedLocation(to) {
-  return to.split('#')[0];
-}